UNSTABLE: All lines of 'u' end with a semicolon
Because tokens are no longer guaranteed to be whitespace delimited
this introduced a very messy, very imperative rewrite of the lex analyzer
that needs some refactoring.

Lex/Syntax error tests are all currently failing and need rework.

STDOUT is no longer a 'sink' and is now a unary operator. In theory
it can be embedded in the middle of a line instead of at the end of
a line, which requires test coverage too.